Raise() the LogWindow after calling Show().  Otherwise the log window might be obscured and the messages missed.  Define a SHOWLOGWINDOW macro to wrap this.
Only call MakeRelativeTo() if we have an absolute path.
removeBaseOdfPath() may not need to calculate a relative path if
the filename does not start with a path separator.

Log any files that cannot be found when reading a .organ file and
mention that they will be excluded/deleted.

When adding a Wave tremulant and only a single ignore Wave tremulant
attack exists, set that attack to play when Wave tremulant is off.
Also log that this has been done.  This behaviour is enabled when
the bool setSingle is true (the default).